adjusts to javascript generator
