+++ generators/javascript/generator.ts
@@ -17,9 +17,11 @@
  * limitations under the License.
  */
 
+import { packageJson } from '../../lib/index.ts';
 import BaseApplicationGenerator from '../base-application/index.ts';
 import BaseSimpleApplicationGenerator from '../base-simple-application/index.ts';
 
+import { isReservedTypescriptKeyword } from './support/reserved-words.ts';
 import type {
   Application as JavascriptApplication,
   Config as JavascriptConfig,
@@ -49,3 +51,118 @@ export class JavascriptApplicationGenerator extends BaseApplicationGenerator<
   JavascriptOptions,
   JavascriptSource
 > {}
+
+export default class JavascriptGenerator extends JavascriptApplicationGenerator {
+  async beforeQueue() {
+    if (!this.fromBlueprint) {
+      await this.composeWithBlueprints();
+    }
+
+    if (!this.delegateToBlueprint) {
+      await this.dependsOnBootstrap('javascript');
+    }
+  }
+
+  get preparingEachEntity() {
+    return this.asPreparingEachEntityTaskGroup({
+      preparing({ entityName }) {
+        if (isReservedTypescriptKeyword(entityName)) {
+          throw new Error(`The entity name "${entityName}" is a reserved TypeScript keyword. It may cause issues in your application.`);
+        }
+      },
+    });
+  }
+
+  get [JavascriptApplicationGenerator.PREPARING_EACH_ENTITY]() {
+    return this.delegateTasksToBlueprint(() => this.preparingEachEntity);
+  }
+
+  get preparingEachEntityField() {
+    return this.asPreparingEachEntityFieldTaskGroup({
+      preparing({ entity, field }) {
+        if (isReservedTypescriptKeyword(field.fieldName)) {
+          throw new Error(`The field name "${field.fieldName}" in entity "${entity.name}" is a reserved TypeScript keyword.`);
+        }
+      },
+    });
+  }
+
+  get [JavascriptApplicationGenerator.PREPARING_EACH_ENTITY_FIELD]() {
+    return this.delegateTasksToBlueprint(() => this.preparingEachEntityField);
+  }
+
+  get preparingEachEntityRelationship() {
+    return this.asPreparingEachEntityRelationshipTaskGroup({
+      preparing({ entity, relationship }) {
+        if (isReservedTypescriptKeyword(relationship.relationshipName)) {
+          throw new Error(
+            `The relationship name "${relationship.relationshipName}" in entity "${entity.name}" is a reserved TypeScript keyword.`,
+          );
+        }
+      },
+    });
+  }
+
+  get [JavascriptApplicationGenerator.PREPARING_EACH_ENTITY_RELATIONSHIP]() {
+    return this.delegateTasksToBlueprint(() => this.preparingEachEntityRelationship);
+  }
+
+  get writing() {
+    return this.asWritingTaskGroup({
+      async writing({ application }) {
+        await this.writeFiles({
+          blocks: [{ templates: [{ override: false, file: 'package.json' }] }],
+          context: application,
+        });
+      },
+    });
+  }
+
+  get [JavascriptApplicationGenerator.WRITING]() {
+    return this.delegateTasksToBlueprint(() => this.writing);
+  }
+
+  get postWriting() {
+    return this.asPostWritingTaskGroup({
+      mergePackageJson({ application, source }) {
+        const {
+          packageJsonNodeEngine,
+          packageJsonType,
+          dasherizedBaseName,
+          projectDescription,
+          packageJsonScripts,
+          clientPackageJsonScripts,
+        } = application;
+
+        this.packageJson.merge({ scripts: packageJsonScripts! });
+
+        this.packageJson.defaults({
+          name: dasherizedBaseName,
+          version: '0.0.0',
+          description: projectDescription,
+          license: 'UNLICENSED',
+        });
+
+        if (packageJsonType === 'module') {
+          this.packageJson.merge({ type: packageJsonType });
+        }
+
+        if (packageJsonNodeEngine) {
+          const packageJsonEngines: any = this.packageJson.get('engines') ?? {};
+          this.packageJson.set('engines', {
+            ...packageJsonEngines,
+            node: typeof packageJsonNodeEngine === 'string' ? packageJsonNodeEngine : packageJson.engines.node,
+          });
+        }
+
+        if (clientPackageJsonScripts && Object.keys(clientPackageJsonScripts).length > 0) {
+          source.mergeClientPackageJson!({ scripts: clientPackageJsonScripts });
+        }
+      },
+    });
+  }
+
+  get [JavascriptApplicationGenerator.POST_WRITING]() {
+    return this.delegateTasksToBlueprint(() => this.postWriting);
+  }
+}
