+++ generators/javascript/templates/package.json.ejs
